# src/main/resources/interfaces/openapi.yaml
openapi: 3.0.1
info:
  title: Realtime balance transaction
  description: Realtime balance calculation service
  version: 1.0.0
paths:
  /balance/v1/health:
    get:
      summary: health check for balance service
      tags:
        - BalanceApi
      operationId: healthCheck
      responses:
        '200':
          description: health check for balance service successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "OK"
        '400':
          description: Invalid input
        '500':
          description: Service exception
  /balance/v1/transaction:
    post:
      summary: Financial transaction between two accounts
      tags:
        - BalanceApi
      operationId: balanceTransaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionRequest'
      responses:
        '200':
          description: Financial transaction successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '400':
          description: Invalid input
        '500':
          description: Service exception
  /balance/v1/query-account:
    get:
      summary: Query financial by account
      tags:
        - BalanceApi
      operationId: queryBalance
      parameters:
        - name: account
          in: query
          schema:
            type: string
          required: true
          description: account
      responses:
        '200':
          description: Query financial by account successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountBalance'
        '400':
          description: Invalid input
        '500':
          description: Service exception
  /balance/v1/init-account:
    post:
      summary: This interface is used to init account data for testing
      tags:
        - BalanceApi
      operationId: initBalance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountBalance'
      responses:
        '200':
          description: This interface is used to init account data for testing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountBalance'
        '400':
          description: Invalid input
        '500':
          description: Service exception
components:
  schemas:
    TransactionRequest:
      type: object
      required:
        - sourceAccount
        - destAccount
        - amount
      properties:
        sourceAccount:
          type: string
          minLength: 1
          maxLength: 100
        destAccount:
          type: string
          minLength: 1
          maxLength: 100
        amount:
          type: number
          format: double
    TransactionResponse:
      type: object
      required:
        - transactionId
        - sourceAccount
        - destAccount
        - amount
        - timestamp
      properties:
        transactionId:
          type: string
          minLength: 1
          maxLength: 100
        sourceAccount:
          type: string
          minLength: 1
          maxLength: 100
        destAccount:
          type: string
          minLength: 1
          maxLength: 100
        amount:
          type: number
          format: double
        timestamp:
          type: string
          minLength: 1
          maxLength: 100
    AccountBalance:
      type: object
      properties:
        account:
          type: string
          minLength: 1
          maxLength: 100
        balance:
          type: number
          format: double